% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_time_series_compare.R
\name{helper_time_series_compare}
\alias{helper_time_series_compare}
\title{Designed for the CM SAF R Toolbox.}
\usage{
helper_time_series_compare(visualizeVariables)
}
\arguments{
\item{visualizeVariables}{A data frame containing all meta data for the plotting process (data.frame).}
}
\description{
This function is a helper function for render_plot_time_series_compare.
}
