% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_hist_plot.R
\name{render_hist_plot}
\alias{render_hist_plot}
\title{Creating a simple histogram.}
\usage{
render_hist_plot(
  dastat,
  shortDescription,
  grid_col,
  bordercolor,
  linesize,
  xlab
)
}
\arguments{
\item{dastat}{Statistics given to hist plot (numeric).}

\item{shortDescription}{A title will be generated using "Histogram of" + description (character).}

\item{grid_col}{Color used for the grid.}

\item{bordercolor}{Color used for borders.}

\item{linesize}{Line width to be used (positive numeric).}

\item{xlab}{Label for x axis (character).}
}
\description{
This routine was implemented for creating histograms in the CM SAF R Toolbox.
}
