% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plot_hovmoller.R
\name{render_plot_hovmoller}
\alias{render_plot_hovmoller}
\title{Plotting routine designed for the CM SAF R Toolbox.}
\usage{
render_plot_hovmoller(
  outfile = NULL,
  fileExtension = ".png",
  visualizeVariables,
  imagewidth,
  imageheight,
  textsize,
  linesize,
  title_data1,
  title_data2,
  nc = NULL,
  nc2 = NULL
)
}
\arguments{
\item{outfile}{Name of the outfile (NULL or character). Should match the fileExtension.
If NULL is passed a file is created in the R session temporary directory.}

\item{fileExtension}{The file extension of the image (character). Has to be one of the following: 'png', 'jpg', 'tif', 'kml', 'pdf'.}

\item{visualizeVariables}{A data frame containing all meta data for the plotting process (data.frame).}

\item{imagewidth}{Width of the image (numeric).}

\item{imageheight}{Height of the image (numeric).}

\item{textsize}{Textsize to be used (cex).}

\item{linesize}{Line width to be used (positive numeric).}

\item{title_data1}{Title of the first data set}

\item{title_data2}{Title of the second data set}

\item{nc}{Alternatively to \code{infile1} you can specify the input as an
object of class \code{ncdf4} (as returned from \code{ncdf4::nc_open}).}

\item{nc2}{Alternatively to \code{infile2} you can specify the input as an
object of class \code{ncdf4} (as returned from \code{ncdf4::nc_open}).}
}
\description{
This function renders a hovmoller plot of two variables.
}
