% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_preview_plot.R
\name{render_preview_plot}
\alias{render_preview_plot}
\title{Creates a preview plot of a selected area}
\usage{
render_preview_plot(spatial_lon_range, spatial_lat_range, lonRange, latRange)
}
\arguments{
\item{spatial_lon_range}{Array containing two values for longitude min and max (numeric).}

\item{spatial_lat_range}{Array containing two values for latitude min and max (numeric).}

\item{lonRange}{Array containing two values for longitude min and max (numeric).}

\item{latRange}{Array containing two values for latitude min and max (numeric).}
}
\description{
This function creates a simple preview plot via maps::map of a given region.
}
