% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmsafvis.R
\docType{package}
\name{cmsafvis-package}
\alias{cmsafvis}
\alias{cmsafvis-package}
\title{cmsafvis: A 'cmsaf' package extension for visualization of CM SAF NetCDF data.}
\description{
The 'cmsafvis' plotting routines are designed to analyze climate files
by generating graphics or videos. The functions (Fieldmean, Fieldmean and anomaly plots) were designed
and tested for daily valued CM SAF NetCDF data. The functions (Absolute Map, Anomaly Map, Climatology,
Warming Stripes Plot, Time Series Plot, Trend Plot) were designed
and tested for daily or monthly valued CM SAF NetCDF data. As interface to NetCDF
data the \link[ncdf4:ncdf4-package]{ncdf4 package} is used.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/cmsaf/cmsaf-r-tools/issues}
}

}
\author{
\strong{Maintainer}: Steffen Kothe \email{Steffen.Kothe@dwd.de}

Other contributors:
\itemize{
  \item Danny Parsons \email{danny@idems.international} [contributor]
}

}
