% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_instat_plot.R
\name{render_instat_plot}
\alias{render_instat_plot}
\title{Creating a simple instat plot.}
\usage{
render_instat_plot(
  co.data,
  shortDescription,
  grid_col,
  bordercolor,
  linesize,
  ylab
)
}
\arguments{
\item{co.data}{Statistics given to hist plot (data.frame).}

\item{shortDescription}{A title will be generated using "Comparison of" + description (character).}

\item{grid_col}{Color used for the grid.}

\item{bordercolor}{Color used for borders.}

\item{linesize}{Line width to be used (positive numeric).}

\item{ylab}{Label for y axis (character).}
}
\description{
This function creates a simple r-instat plot.
}
